+++ HISTORY.md
@@ -1,6 +1,8 @@
 ## v1.0.1
 * Added Cloudflare Turnstile `TestingSiteKeys` and `TestingSecretKeys`.
 * Changed `ClientResponseKey` -> `ClientResponseFormKey`.
+* VerifyAsync method signature changed; made remoteIp param optional per documnetation.
+* Support `CData` field in `TurnstileResponse`.
 
 ## v1.0.0
 * Initial release.